requirements.txt
google-adk==1.19.0
pandas==2.3.3
python-dotenv==1.2.1
google-adk[a2a]
